+++ src/main/java/chess/domain/BoardFactory.java
@@ -1,5 +1,16 @@
 package chess.domain;
 
+import chess.domain.color.Color;
+import chess.domain.piece.Bishop;
+import chess.domain.piece.BlackFirstPawn;
+import chess.domain.piece.Blank;
+import chess.domain.piece.King;
+import chess.domain.piece.Knight;
+import chess.domain.piece.Piece;
+import chess.domain.piece.Position;
+import chess.domain.piece.Queen;
+import chess.domain.piece.Rook;
+import chess.domain.piece.WhiteFirstPawn;
 import java.util.HashMap;
 import java.util.Map;
