+++ src/main/java/chess/domain/color/Color.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.color;
 
 public enum Color {
     WHITE,
