+++ src/main/java/chess/domain/piece/BlackPawn.java
@@ -1,11 +1,12 @@
-package chess.domain;
+package chess.domain.piece;
 
+import chess.domain.color.Color;
 import java.util.Set;
 
-public class BlackFirstPawn extends Pawn {
-    private static Set<Direction> DIRECTIONS = Direction.getBlackFirstPawnDirection();
+public class BlackPawn extends Pawn {
+    private static Set<Direction> DIRECTIONS = Set.of(Direction.DOWN, Direction.LEFT_DOWN, Direction.RIGHT_DOWN);
 
-    public BlackFirstPawn(Position position) {
+    public BlackPawn(Position position) {
         super(position, Color.BLACK, DIRECTIONS);
     }
