+++ src/main/java/chess/domain/piece/Blank.java
@@ -1,5 +1,6 @@
-package chess.domain;
+package chess.domain.piece;
 
+import chess.domain.color.Color;
 import java.util.Set;
 
 public class Blank extends Piece {
