+++ src/main/java/chess/domain/piece/WhitePawn.java
@@ -1,11 +1,12 @@
-package chess.domain;
+package chess.domain.piece;
 
+import chess.domain.color.Color;
 import java.util.Set;
 
-public class WhiteFirstPawn extends Pawn {
-    private static Set<Direction> DIRECTIONS = Direction.getWhiteFirstPawnDirection();
+public class WhitePawn extends Pawn {
+    private static Set<Direction> DIRECTIONS = Set.of(Direction.UP, Direction.LEFT_UP, Direction.RIGHT_UP);
 
-    public WhiteFirstPawn(Position position) {
+    public WhitePawn(Position position) {
         super(position, Color.WHITE, DIRECTIONS);
     }
