+++ src/main/java/chess/domain/strategy/MoveStrategy.java
@@ -1,10 +1,10 @@
 package chess.domain.strategy;
 
-import chess.domain.Blank;
-import chess.domain.Color;
-import chess.domain.Piece;
-import chess.domain.PieceType;
-import chess.domain.Position;
+import chess.domain.piece.Blank;
+import chess.domain.color.Color;
+import chess.domain.piece.Piece;
+import chess.domain.piece.PieceType;
+import chess.domain.piece.Position;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
