+++ src/main/java/chess/domain/strategy/PawnMoveStrategy.java
@@ -1,11 +1,11 @@
 package chess.domain.strategy;
 
-import chess.domain.Blank;
-import chess.domain.Color;
-import chess.domain.Pawn;
-import chess.domain.Piece;
-import chess.domain.PieceType;
-import chess.domain.Position;
+import chess.domain.piece.Blank;
+import chess.domain.color.Color;
+import chess.domain.piece.Pawn;
+import chess.domain.piece.Piece;
+import chess.domain.piece.PieceType;
+import chess.domain.piece.Position;
 import java.util.Map;
 import java.util.Set;
