+++ src/test/java/chess/domain/TestBoardFactory.java
@@ -1,5 +1,8 @@
 package chess.domain;
 
+import chess.domain.piece.Blank;
+import chess.domain.piece.Piece;
+import chess.domain.piece.Position;
 import java.util.HashMap;
 import java.util.Map;
