+++ src/test/java/chess/domain/piece/BlackPawnTest.java
@@ -1,9 +1,10 @@
-package chess.domain;
+package chess.domain.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
+import chess.domain.color.Color;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
