+++ src/test/java/chess/domain/strategy/PawnMoveStrategyTest.java
@@ -5,9 +5,9 @@ import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException
 import static org.junit.jupiter.api.Assertions.*;
 
 import chess.domain.BoardFactory;
-import chess.domain.Color;
-import chess.domain.PieceType;
-import chess.domain.Position;
+import chess.domain.color.Color;
+import chess.domain.piece.PieceType;
+import chess.domain.piece.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
